Fixed npm run build issue
npm run build would fail inside of docker and work outside complaining that the methone package wasn't installed. The methone package was installed via a github repo. I investigated how the package was installed differently inside and outside docker and noticed that the dist directory was missing inside the docker image and was present outside. The dist directory is generated when methone's prepare script is run during npm install. But apparently the prepare script won't run as root (https://stackoverflow.com/questions/52766731/npm-v6-4-1-not-running-prepare-inside-docker). So I made the dockerfile switch to a normal user before running npm install and now it works.